chore(reorg): Move events modal folder to the `src/events` folder
